+++ CHANGELOG.md
@@ -28,6 +28,7 @@ An AssertJ style API for testing KtLint rules ([#1444](https://github.com/pinter
 - Fix formatting of a property delegate with a dot-qualified-expression `indent` ([#1340](https://github.com/pinterest/ktlint/ssues/1340))
 - Keep formatting of for-loop in sync with default IntelliJ formatter (`indent`) and a newline in the expression in a for-statement should not force to wrap it `wrapping` ([#1350](https://github.com/pinterest/ktlint/issues/1350))
 - Fix indentation of property getter/setter when the property has an initializer on a separate line `indent` ([#1335](https://github.com/pinterest/ktlint/issues/1335))
+- Add kotlinx to Idea configuration for ImportOrderingRule
 
 ### Changed
 - Update Kotlin development version to `1.7.0-RC` and Kotlin version to `1.6.21`.